Fixed unused maxiters option
The option `maxiters`, stored as `simu_opts->maxiters` was not used after storing it, and the number of times the loop runs was hard coded to be 20. Also changed the option-descriptions to better reflect the meaning of the `maxiters` option, which, if I understand correctly, is the number of times the outer loop runs while the inner loop is fixed to 1000 steps/iterations for each of the outer loop's run.